# [Opt](Vectorized) Support push down no grouping agg