--- a/be/src/olap/rowset/segment_v2/segment.cpp
+++ b/be/src/olap/rowset/segment_v2/segment.cpp
@@ -36,6 +36,8 @@
 #include "util/crc32c.h"
 #include "util/slice.h" // Slice
 
+#include "vec/olap/vgeneric_iterators.h"
+
 namespace doris {
 namespace segment_v2 {
 
@@ -100,7 +102,11 @@ Status Segment::new_iterator(const Schema& schema, const StorageReadOptions& rea
     }
 
     RETURN_IF_ERROR(load_index());
-    iter->reset(new SegmentIterator(this->shared_from_this(), schema));
+    if (read_options.col_id_to_del_predicates.empty() && read_options.push_down_agg_type_opt != TPushAggOp::NONE) {
+        iter->reset(vectorized::new_vstatistics_iterator(this->shared_from_this(), schema));
+    } else {
+        iter->reset(new SegmentIterator(this->shared_from_this(), schema));
+    }
     iter->get()->init(read_options);
     return Status::OK();
 }