--- a/be/src/vec/exec/scan/new_olap_scanner.cpp
+++ b/be/src/vec/exec/scan/new_olap_scanner.cpp
@@ -152,8 +152,9 @@ Status NewOlapScanner::_init_tablet_reader_params(
         _tablet_reader_params.direct_mode = true;
         _aggregation = true;
     } else {
-        _tablet_reader_params.direct_mode = _aggregation || single_version ||
-                                     real_parent->_olap_scan_node.__isset.push_down_agg_type_opt;
+        _tablet_reader_params.direct_mode =
+                _aggregation || single_version ||
+                real_parent->_olap_scan_node.__isset.push_down_agg_type_opt;
     }
 
     _tablet_reader_params.tablet = _tablet;
@@ -161,7 +162,8 @@ Status NewOlapScanner::_init_tablet_reader_params(
     _tablet_reader_params.reader_type = READER_QUERY;
     _tablet_reader_params.aggregation = _aggregation;
     if (real_parent->_olap_scan_node.__isset.push_down_agg_type_opt)
-        _tablet_reader_params.push_down_agg_type_opt = real_parent->_olap_scan_node.push_down_agg_type_opt;
+        _tablet_reader_params.push_down_agg_type_opt =
+                real_parent->_olap_scan_node.push_down_agg_type_opt;
     _tablet_reader_params.version = Version(0, _version);
 
     // Condition