--- a/be/src/vec/exec/volap_scanner.cpp
+++ b/be/src/vec/exec/volap_scanner.cpp
@@ -185,7 +185,8 @@ Status VOlapScanner::_init_tablet_reader_params(
     _tablet_reader_params.reader_type = READER_QUERY;
     _tablet_reader_params.aggregation = _aggregation;
     if (_parent->_olap_scan_node.__isset.push_down_agg_type_opt)
-        _tablet_reader_params.push_down_agg_type_opt = _parent->_olap_scan_node.push_down_agg_type_opt;
+        _tablet_reader_params.push_down_agg_type_opt =
+                _parent->_olap_scan_node.push_down_agg_type_opt;
     _tablet_reader_params.version = Version(0, _version);
 
     // Condition