--- a/be/src/vec/olap/vgeneric_iterators.cpp
+++ b/be/src/vec/olap/vgeneric_iterators.cpp
@@ -21,6 +21,8 @@
 
 #include "common/status.h"
 #include "olap/iterators.h"
+#include "olap/rowset/segment_v2/column_reader.h"
+#include "olap/rowset/segment_v2/segment.h"
 #include "olap/schema.h"
 #include "vec/core/block.h"
 
@@ -110,6 +112,78 @@ private:
 Status VAutoIncrementIterator::init(const StorageReadOptions& opts) {
     return Status::OK();
 }
+
+class VStatisticsIterator : public RowwiseIterator {
+public:
+    // Will generate num_rows rows in total
+    VStatisticsIterator(std::shared_ptr<Segment> segment, const Schema& schema)
+            : _segment(std::move(segment)), _schema(schema) {}
+
+    ~VStatisticsIterator() override {
+        for (auto& pair : _column_iterators_map) {
+            delete pair.second;
+        }
+    }
+
+    Status init(const StorageReadOptions& opts) override {
+        if (!_init) {
+            _push_down_agg_type_opt = opts.push_down_agg_type_opt;
+
+            for (size_t i = 0; i < _schema.num_column_ids(); i++) {
+                auto cid = _schema.column_id(i);
+                auto unique_id = _schema.column(cid)->unique_id();
+                if (_column_iterators_map.count(unique_id) < 1) {
+                    RETURN_IF_ERROR(_segment->new_column_iterator(opts.tablet_schema->column(cid),
+                                                                  &_column_iterators_map[unique_id]));
+                }
+                _column_iterators.push_back(_column_iterators_map[unique_id]);
+            }
+
+            _target_rows =  _push_down_agg_type_opt == TPushAggOp::MINMAX ? 2 : _segment->num_rows();
+            _init = true;
+        }
+
+        return Status::OK();
+    }
+
+    Status next_batch(Block* block) override {
+        DCHECK(block->columns() == _column_iterators.size());
+        if (_output_rows < _target_rows) {
+            block->clear_column_data();
+            auto columns = block->mutate_columns();
+
+            size_t size = _push_down_agg_type_opt == TPushAggOp::MINMAX ?
+                 2 : std::min(_target_rows - _output_rows, MAX_ROW_SIZE_IN_COUNT);
+            if (_push_down_agg_type_opt == TPushAggOp::COUNT) {
+                size = std::min(_target_rows - _output_rows, MAX_ROW_SIZE_IN_COUNT);
+                for (int i = 0; i < block->columns(); ++i) {
+                    columns[i]->resize(size);
+                }
+            } else {
+                for (int i = 0; i < block->columns(); ++i) {
+                    _column_iterators[i]->next_batch_of_zone_map(&size, columns[i]);
+                }
+            }
+            _output_rows += size;
+            return Status::OK();
+        }
+        return Status::EndOfFile("End of VStatisticsIterator");
+    }
+
+    const Schema& schema() const override { return _schema; }
+
+private:
+    std::shared_ptr<Segment> _segment;
+    const Schema& _schema;
+    size_t _target_rows = 0;
+    size_t _output_rows = 0;
+    bool _init = false;
+    TPushAggOp::type _push_down_agg_type_opt;
+    std::map<int32_t, ColumnIterator*> _column_iterators_map;
+    std::vector<ColumnIterator*> _column_iterators;
+
+    static constexpr size_t MAX_ROW_SIZE_IN_COUNT = 65535;
+};
 
 // Used to store merge state for a VMergeIterator input.
 // This class will iterate all data from internal iterator
@@ -573,6 +647,10 @@ RowwiseIterator* new_union_iterator(std::vector<RowwiseIterator*>& inputs) {
         return *(inputs.begin());
     }
     return new VUnionIterator(inputs);
+}
+
+RowwiseIterator* new_vstatistics_iterator(std::shared_ptr<Segment> segment, const Schema& schema) {
+    return new VStatisticsIterator(segment, schema);
 }
 
 RowwiseIterator* new_auto_increment_iterator(const Schema& schema, size_t num_rows) {