--- a/be/src/vec/olap/vgeneric_iterators.h
+++ b/be/src/vec/olap/vgeneric_iterators.h
@@ -43,6 +43,8 @@ RowwiseIterator* new_union_iterator(std::vector<RowwiseIterator*>& inputs);
 // Client should delete returned iterator.
 RowwiseIterator* new_auto_increment_iterator(const Schema& schema, size_t num_rows);
 
+RowwiseIterator* new_vstatistics_iterator(std::shared_ptr<Segment> segment, const Schema& schema);
+
 } // namespace vectorized
 
 } // namespace doris