--- a/be/src/vec/olap/vgeneric_iterators.h
+++ b/be/src/vec/olap/vgeneric_iterators.h
@@ -19,6 +19,10 @@
 
 namespace doris {
 
+namespace segment_v2 {
+class Segment;
+}
+
 namespace vectorized {
 
 // Create a merge iterator for input iterators. Merge iterator will merge