… fe/fe-core/src/main/java/org/apache/doris/qe/SessionVariable.java
… gensrc/thrift/PlanNodes.thrift